Move supporting code into separate file
