+++ tests/checks.rs
@@ -1,89 +1,9 @@
-use std::{fmt::Debug, path::PathBuf};
-
-use anyhow::bail;
 use rstest::{fixture, rstest};
+use std::path::PathBuf;
+use test_utils::TestDir as TD;
 use version_control_clean_check::{check_version_control, CheckOptions, VCSError, VCSResult};
-use TestDir as TD;
 
-#[derive(Debug)]
-struct TestError(VCSError);
-impl From<VCSError> for TestError {
-    fn from(value: VCSError) -> Self {
-        Self(value)
-    }
-}
-
-impl<T: Debug> TryFrom<VCSResult<T>> for TestError {
-    type Error = anyhow::Error;
-
-    fn try_from(value: VCSResult<T>) -> Result<Self, Self::Error> {
-        if value.is_ok() {
-            bail!("Value is not error. Found: {value:?}");
-        }
-        Ok(Self(value.unwrap_err()))
-    }
-}
-
-impl PartialEq for TestError {
-    fn eq(&self, other: &Self) -> bool {
-        match (&self.0, &other.0) {
-            (VCSError::NoVCS, VCSError::NoVCS) => true,
-            (
-                VCSError::NotAllowedFilesFound {
-                    dirty_files: l_dirty_files,
-                    staged_files: l_staged_files,
-                },
-                VCSError::NotAllowedFilesFound {
-                    dirty_files: r_dirty_files,
-                    staged_files: r_staged_files,
-                },
-            ) => l_dirty_files == r_dirty_files && l_staged_files == r_staged_files,
-            (VCSError::GitError(..), _) | (VCSError::Anyhow(..), _) => false, // Never equal if not one of our local errors during testing
-            _ => core::mem::discriminant(&self.0) == core::mem::discriminant(&other.0),
-        }
-    }
-}
-
-enum TestDir {
-    NoVCS,
-    Clean,
-    StagedOnly,
-    DirtyOnly,
-    StagedAndDirty,
-}
-
-impl TestDir {
-    const TEST_DIR_BASE: &str = "tests/test_folders/";
-    fn to_path(&self) -> PathBuf {
-        let base_test_folder = PathBuf::from(Self::TEST_DIR_BASE);
-        let sub_folder = match self {
-            TestDir::NoVCS => "no_vcs",
-            TestDir::Clean => "clean",
-            TestDir::StagedOnly => "staged_only",
-            TestDir::DirtyOnly => "dirty_only",
-            TestDir::StagedAndDirty => "staged_and_dirty",
-        };
-        let result = base_test_folder.join(sub_folder);
-        assert!(result.exists(), "Path not found: {result:?}");
-        result.canonicalize().unwrap()
-    }
-}
-
-fn match_results(actual: VCSResult<()>, expected: VCSResult<()>) {
-    match (&actual, &expected) {
-        (Ok(_), Ok(_)) => (),
-        (Ok(_), Err(_)) | (Err(_), Ok(_)) => {
-            panic!("Actual and Expected do not match./n actual:{actual:?}/n expected: {expected:?}")
-        }
-        (Err(..), Err(..)) => {
-            let actual_error = actual.unwrap_err();
-            let expected_error = expected.unwrap_err();
-            println!("---\nActual Error:\n{actual_error}\n");
-            println!("---\nExpected Error:\n{expected_error}\n---");
-            assert_eq!(TestError(actual_error), TestError(expected_error))
-        }
-    }
-}
+mod test_utils;
 
 #[test]
 fn non_existent_folder() {
@@ -93,19 +13,12 @@ fn non_existent_folder() {
 
     // Test is no vcs
     let actual = check_version_control(&non_existent_path, &opts);
-    match_results(actual, Err(VCSError::NoVCS));
+    test_utils::match_results(actual, Err(VCSError::NoVCS));
 
     // Test passes if no vcs allowed
     opts.allow_no_vcs = true;
     let actual = check_version_control(&non_existent_path, &opts);
-    match_results(actual, Ok(()));
-}
-
-fn test_check_version_control(opts: CheckOptions, test_dir: TestDir, expected: VCSResult<()>) {
-    let path = test_dir.to_path();
-    println!("Opts: {opts:#?}\nPath: {path:?}");
-    let actual = check_version_control(path, &opts);
-    match_results(actual, expected);
+    test_utils::match_results(actual, Ok(()));
 }
 
 #[fixture]
@@ -121,12 +34,12 @@ fn create_dirs() -> anyhow::Result<()> {
 #[case(TD::StagedOnly)]
 #[case(TD::DirtyOnly)]
 #[case(TD::StagedAndDirty)]
-fn allow_no_vcs(#[case] test_dir: TestDir, create_dirs: &anyhow::Result<()>) {
+fn allow_no_vcs(#[case] test_dir: test_utils::TestDir, create_dirs: &anyhow::Result<()>) {
     assert!(create_dirs.is_ok(), "{create_dirs:?}");
     let mut opts = CheckOptions::new();
     opts.allow_no_vcs = true;
     let expected = Ok(());
-    test_check_version_control(opts, test_dir, expected);
+    test_utils::test_check_version_control(opts, test_dir, expected);
 }
 
 #[rstest]
@@ -157,7 +70,7 @@ fn allow_no_vcs(#[case] test_dir: TestDir, create_dirs: &anyhow::Result<()>) {
 fn vcs_required(
     #[case] allow_dirty: bool,
     #[case] allow_staged: bool,
-    #[case] test_dir: TestDir,
+    #[case] test_dir: test_utils::TestDir,
     #[case] expected: VCSResult<()>,
     create_dirs: &anyhow::Result<()>,
 ) {
@@ -167,5 +80,5 @@ fn vcs_required(
     opts.allow_dirty = allow_dirty;
     opts.allow_staged = allow_staged;
 
-    test_check_version_control(opts, test_dir, expected);
+    test_utils::test_check_version_control(opts, test_dir, expected);
 }
